requirements.txt
django==2.2.6
pillow==6.2.0
dj-database-url==0.5.0
psycopg2==2.8.4
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
